backend/ai_modules/video_analyzer.py
"""
Video Analysis Module - Placeholder Implementation
Detects baby presence and activity (sleeping/sitting).
Replace with production models (MediaPipe, CNN, ResNet) later.
"""

import cv2
import numpy as np
from typing import Dict, Tuple

class VideoAnalyzer:
    def __init__(self):
        """Initialize placeholder video analyzer."""
        self.face_cascade = cv2.CascadeClassifier(
            cv2.data.haarcascades + 'haarcascade_frontalface_default.xml'
        )
        self.motion_threshold = 5.0

    def analyze(self, video_path: str) -> Dict:
        """
        Analyze video file for baby presence and activity.

        Args:
            video_path: Path to video file

        Returns:
            Dict with presence, activity, and confidence
        """
        try:
            cap = cv2.VideoCapture(video_path)

            if not cap.isOpened():
                return self._no_presence_result("Could not open video")

            frames = []
            frame_count = 0
            max_frames = 120  # ~4 seconds at 30fps

            # Read frames
            while frame_count < max_frames:
                ret, frame = cap.read()
                if not ret:
                    break
                frames.append(frame)
                frame_count += 1

            cap.release()

            if len(frames) < 10:
                return self._no_presence_result("Insufficient frames")

            # Detect presence
            has_presence, presence_confidence = self._detect_presence(frames)

            if not has_presence:
                return {
                    "presence": False,
                    "activity": None,
                    "confidence": float(presence_confidence)
                }

            # Classify activity
            activity, activity_confidence = self._classify_activity(frames)

            return {
                "presence": True,
                "activity": activity,
                "confidence": float(activity_confidence)
            }

        except Exception as e:
            print(f"Error analyzing video: {e}")
            return self._no_presence_result(str(e))

    def _no_presence_result(self, reason: str = None) -> Dict:
        """Return standard no presence result."""
        return {
            "presence": False,
            "activity": None,
            "confidence": 0.85,
            "reason": reason
        }

    def _detect_presence(self, frames: list) -> Tuple[bool, float]:
        """
        Placeholder presence detection using face detection.

        Replace with trained model:
        - Use MediaPipe Face/Pose detection
        - Run baby detection CNN
        - Check for baby-specific features (size, proportions)
        """
        face_count = 0
        total_checked = 0

        # Check every 5th frame for efficiency
        for i in range(0, len(frames), 5):
            frame = frames[i]
            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)

            # Detect faces
            faces = self.face_cascade.detectMultiScale(
                gray,
                scaleFactor=1.1,
                minNeighbors=5,
                minSize=(30, 30)
            )

            if len(faces) > 0:
                face_count += 1
            total_checked += 1

        # If faces detected in >30% of checked frames, consider baby present
        presence_ratio = face_count / total_checked if total_checked > 0 else 0
        has_presence = presence_ratio > 0.3

        confidence = presence_ratio if has_presence else (1.0 - presence_ratio)
        confidence = min(max(confidence, 0.6), 0.95)

        return has_presence, confidence

    def _classify_activity(self, frames: list) -> Tuple[str, float]:
        """
        Placeholder activity classification.

        Replace with trained classifier:
        - Use MediaPipe Pose landmarks
        - Extract keypoint positions (eyes, limbs, torso)
        - Classify: sleeping (closed eyes, horizontal), sitting (upright pose)
        """
        # Calculate motion between frames
        motion_scores = []

        for i in range(1, min(len(frames), 50)):
            prev_gray = cv2.cvtColor(frames[i - 1], cv2.COLOR_BGR2GRAY)
            curr_gray = cv2.cvtColor(frames[i], cv2.COLOR_BGR2GRAY)

            # Calculate frame difference
            diff = cv2.absdiff(prev_gray, curr_gray)
            motion_score = np.mean(diff)
            motion_scores.append(motion_score)

        avg_motion = np.mean(motion_scores) if motion_scores else 0

        # Simple heuristic: low motion = sleeping, high motion = sitting/active
        if avg_motion < self.motion_threshold:
            activity = "sleeping"
            confidence = min(0.75 + (self.motion_threshold - avg_motion) / 20, 0.94)
        else:
            activity = "sitting"
            confidence = min(0.70 + (avg_motion - self.motion_threshold) / 20, 0.92)

        return activity, confidence
