backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
opencv-python==4.8.1.78
twilio==8.10.0
supabase==2.3.0
